examples/io_uring-udp: compute buffer size and offset in size_t
buffer_size() and get_buffer() shift by ctx->buf_shift, which comes from
the -b option (log2 of the buffer size), but evaluate 1U << buf_shift and
idx << buf_shift in 32-bit before widening to size_t / a pointer. For a
shift >= 32 that is undefined behavior, and even otherwise the result
truncates to 32 bits, so buffer_size() returns a wrong value and
get_buffer() forms an out-of-range pointer. Building with
-fsanitize=undefined and running with -b 40 reports

  runtime error: shift exponent 40 is too large for 32-bit type

and buffer_size() returns 256 instead of 2^40. Do both shifts in size_t.

+++ examples/io_uring-udp.c
@@ -36,12 +36,12 @@ struct ctx {
 
 static size_t buffer_size(struct ctx *ctx)
 {
-	return 1U << ctx->buf_shift;
+	return (size_t)1 << ctx->buf_shift;
 }
 
 static unsigned char *get_buffer(struct ctx *ctx, int idx)
 {
-	return ctx->buffer_base + (idx << ctx->buf_shift);
+	return ctx->buffer_base + ((size_t)idx << ctx->buf_shift);
 }
 
 static int setup_buffer_pool(struct ctx *ctx)
